FIX: No line is appended to just initialized table
And the corresponding test (t3000-update-initialize.bats) didn't catch that because it reused the table created by a previous test (so only the very first test run would fail).
Skip to :skipHeader label instead of end of sed script in IGNORE_HEADER_LINE_SED_ARGS, and introduce such label at the right position in update() (and at the end in query()).
Add clean_table() test helper and invoke that in t3000-update-initialize.bats.