ENH: Normalize strange schema names read from the table
Spaces previously messed up the parsing (i.e. more variables than columns), and other characters that are not allowed as variable names would cause errors.
Instead, replace any non-alphanumeric non-underscore character with an underscore to obtain kosher column names.
We now cannot use the trick of passing (-a COL) to iterateOverRecords(). Instead, pass explicit iterateReadArgs for the special case (schema is a global array and can be referenced as-is).