ENH: Add -k|--query-keys
Though this can be obtained via --each 'echo $KEY', the variable is dependent on the (used / passed) schema, and the whole is not very expressive. As we can skip iteration and before / after commands for it, I think it's worth having a dedicated option for such a common use case.