Just truncate lock file instead of removing it
Apparently, when flock is waiting on the lock, and the current owner removes it, the lock file isn't re-created, but the wait is over and the critical section entered without creating a lock. So "is created if it does not already exist" seems to mean when entering the wait loop, but not within it.
Anyway, keeping the lock file around also enables a check for non-transactional access (and a warning to the user), so it's not so bad.