BUG: Empty columns are not handled properly
The parsing via IFS=$'\t' skips continuous runs of tabs instead of using exactly one tab as the separator. In Bash 4, we could use mapfile on a line that got tabs translated to newlines, in Bash 4.4 we could even specify the custom separator. Let's do it via a manual loop here to also support Bash 3.