FIX: Rollover must accept leading 9s as well and we cannot immediately increment an 8 to a 9
rollover${c} turns all trailing 9s to 0s, but if the incremented digit was an 8, that would turn to a 9 as well, and the position information (which is hard to pass otherwise in sed) would be lost! Instead, increment the 8 into an intermediate symbol ("_") first, and let rollover${c} convert it while turning the first 9 into a 0. From then on, there's that 0 in between, and any following 9s will be processed correctly.