WIP: Implement output of original line on update
As the sed does inline processing of $TABLE_FILESPEC, we explicitly write the updated line to stdout (which goes into the command substitution).
This is done in the moment when the existing line is recognized, right after copying it into the hold space.
We could also do the printing on the final line, when the hold space is examined, but that clashes with the sed c command in updateFullRecord(), which restarts processing at the next line, and so the $ address won't be reached when the update is on the last line. (Which doesn't matter for the newRecord appending, as there definitely was an existing line when this happens.)
The output formatting is done by outputQueriedLine() for us; we just need to wire it together.